+++ proxystore/p2p/relay/basic/client.py
@@ -35,7 +35,7 @@ from proxystore.p2p.relay.messages import RelayMessageDecodeError
 from proxystore.p2p.relay.messages import RelayRegistrationRequest
 from proxystore.p2p.relay.messages import RelayResponse
 from proxystore.p2p.task import spawn_guarded_background_task
-from proxystore.utils import hostname
+from proxystore.utils.environment import hostname
 
 logger = logging.getLogger(__name__)
