+++ proxystore/utils/data.py
@@ -1,12 +1,8 @@
-"""General purpose utility functions."""
+"""Utilities for interacting with data."""
 from __future__ import annotations
 
 import decimal
-import importlib
-import os
 import re
-import socket
-from typing import Any
 from typing import Generator
 
 
@@ -26,73 +22,6 @@ def chunk_bytes(
     length = len(data)
     for index in range(0, length, chunk_size):
         yield data[index : min(index + chunk_size, length)]
-
-
-def get_class_path(cls: type[Any]) -> str:
-    """Get the fully qualified path of a type.
-
-    Example:
-        ```python
-        >>> from proxystore.connectors.connector import Connector
-        >>> get_class_path(Connector)
-        'proxystore.connectors.connector.Connector'
-        ```
-
-    Args:
-        cls: Class type to get fully qualified path of.
-
-    Returns:
-        Fully qualified path of `cls`.
-    """
-    return f'{cls.__module__}.{cls.__qualname__}'
-
-
-def import_class(path: str) -> type[Any]:
-    """Import class via its fully qualified path.
-
-    Example:
-        ```python
-        >>> import_class('proxystore.connectors.connector.Connector')
-        <class 'proxystore.connectors.connector.Connector'>
-        ```
-
-    Args:
-        path: Fully qualified path of class to import.
-
-    Returns:
-        Imported class.
-
-    Raises:
-        ImportError: If a class at the `path` is not found.
-    """
-    module_path, _, name = path.rpartition('.')
-    if len(module_path) == 0:
-        raise ImportError(
-            f'Class path must contain at least one module. Got {path}',
-        )
-    module = importlib.import_module(module_path)
-    return getattr(module, name)
-
-
-def home_dir() -> str:
-    """Return the absolute path to the proxystore home directory.
-
-    If set, `$PROXYSTORE_HOME` is preferred. Otherwise,
-    `$XDG_DATA_HOME/proxystore` is returned where `$XDG_DATA_HOME` defaults
-    to `$HOME/.local/share` if unset.
-    """
-    path = os.environ.get('PROXYSTORE_HOME')
-    if path is None:
-        prefix = os.environ.get('XDG_DATA_HOME') or os.path.expanduser(
-            '~/.local/share',
-        )
-        path = os.path.join(prefix, 'proxystore')
-    return os.path.abspath(path)
-
-
-def hostname() -> str:
-    """Return current hostname."""
-    return socket.gethostname()
 
 
 def bytes_to_readable(size: int, precision: int = 3) -> str:
